changed double to decimal type

+++ Lab4/Lab4_s26486/Services/InventoryService.cs
@@ -53,7 +53,7 @@ public class InventoryService
         _productWarehouseRepository.Add(productWarehouse);
     }
 
-    private double CalculatePrice(int productId)
+    private decimal CalculatePrice(int productId)
     {
         var product = _productRepository.GetById(productId);
